zhehe/requirements/base.txt
django>=3.0.3
celery>=4.4.0
pandoc>=2.9.2.1
pypandoc>=1.4
psycopg2==2.8.4
django-crispy-forms>=1.9.1
django-allauth==0.42.0